refactor: DetailPost Comment component 분리
