+++ src/pages/Post.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import PostDetail from "../components/Post/PostDetail";
+
 import PostList from "../components/Post/PostList";
 
 const Post = () => {
@@ -12,7 +12,6 @@ const Post = () => {
   return (
     <PostWrapper>
       <PostList />
-      {/* <PostDetail /> */}
     </PostWrapper>
   );
 };
